use getMessage instead of getFeed
