+++ util.js
@@ -38,29 +38,51 @@ function findByPath (indexes, path) {
   })
 }
 
-function isFunction (variable) {
-  return typeof variable === 'function'
-}
-
 function validator (msg) {
   if (typeof msg !== 'object') return null
   if (typeof msg.value !== 'object') return null
   if (typeof msg.value.timestamp !== 'number') return null
-  if (typeof msg.value.type !== 'string') return null
   return msg
 }
 
-function fromMultifeed (feeds) {
-  return function getFeed (msg, cb) {
+function fromMultifeed (feeds, opts = {}) {
+  var validate = opts.validator || validator
+
+  return function getMessage (msg, cb) {
     var msgId = msg.value
     var [ feedId, sequence ] = msgId.split('@')
     var feed = feeds.feed(feedId)
     var seq = Number(sequence)
-    return cb(null, { feed, seq })
+    feed.get(seq, (err, value) => {
+      var msg = validator({
+        key: feed.key.toString('hex'),
+        seq,
+        value
+      })
+      if (!msg) return cb(new Error('message failed to validate'))
+      return cb(null, msg)
+    })
   }
 }
 
-function fromHypercore (feed) {
+function fromHypercore (feed, opts = {}) {
+  var validate = opts.validator || validator
+
+  return function getMessage (msg, cb) {
+    var msgId = msg.value
+    var sequence = msgId.split('@')[1]
+    var seq = Number(sequence)
+
+    feed.get(seq, (err, value) => {
+      var msg = validate({
+        key: feed.key.toString('hex'),
+        seq,
+        value
+      })
+      if (!msg) return cb(new Error('message failed to validate'))
+      return cb(null, msg)
+    })
+  }
 }
 
 module.exports = {
@@ -68,7 +90,7 @@ module.exports = {
   get,
   set,
   findByPath,
-  isFunction,
   validator,
-  fromMultifeed
+  fromMultifeed,
+  fromHypercore
 }
